# public/docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "Taklimakan cryptodata",
    "version": "v0.0.1"
  },
  "paths": {
...
    "/api/top/tn_indexes/charts": {
      "get": {
        "tags": [
          "top"
        ],
        "summary": "Get data for charts TN index",
        "description": "Data charts interval time from to",
        "operationId": "getCharts",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Index type",
            "required": false,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "tn10",
                "tn50",
                "tn100",
                "tn200"
              ]
            },
            "collectionFormat": "multi"
          },
          {
            "name": "period_date_start",
            "in": "query",
            "description": "Date start",
            "required": false,
            "type": "string"
          },
          {
            "name": "period_date_end",
            "in": "query",
            "description": "Date end",
            "required": false,
            "type": "string"
          },
          {
            "name": "object_amount",
            "in": "query",
            "description": "Amount object in response",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name",
                "photoUrls"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "period_date_end": {
                      "type": "string"
                    },
                    "period_interval": {
                      "type": "string"
                    },
                    "object_amount": {
                      "type": "string"
                    }
                  }
                },
                "data": {
                  "type": "array",
                  "xml": {
                    "name": "",
                    "wrapped": true
                  },
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
  },
  "consumes": [
    "application/json"
  ]
}